refactor: 네비게이션 로직 모듈화 및 구조 개선
* refactor: MainActivity의 NavHost 로직을 각 피처 모듈별 navigation 함수로 분리
* add: 각 피처 모듈(`senior`, `onboarding`, `guardian`, `report`)에 Navigation 확장 함수 추가
* refactor: `sharedViewModel` 확장 함수를 `core:ui` 모듈로 이동 및 공통화
* chore: `feature:onboarding` 모듈에 `core:ui` 의존성 추가
