+++ feature/onboarding/build.gradle.kts
@@ -11,6 +11,7 @@ android {
 dependencies {
     implementation(projects.core.designsystem)
     implementation(projects.core.navigation)
+    implementation(projects.core.ui)
     implementation(projects.domain)
 
     implementation(libs.timber)
